requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-ipware==7.0.1
PyYAML==6.0.1
ua-parser==0.18.0
user-agents==2.2.0
django-user-agents==0.4.0